Refactor: Remove unused layout files and update resource references
This commit removes the following unused layout files:
- activity_login_old.xml
- popup_date.xml

It also removes the following unused drawable resources:
- food_category.xml
- transportation_category.xml
- communications_category.png
- insurance_category.png
- utilities_category.png

Additionally, the commit updates resource references in `expense_item.xml` and `category_grid_item.xml` to use `ic_food` instead of the removed `food_category`.
Updated code to get the current user ID.
Modified the back press logic in LoginActivity.
Added the screenOrientation attribute in `AndroidManifest.xml` to set the activity to portrait mode.
Modified the logic in the MainActivity to skip the GetStartedActivity if a user is logged in.